Extract out logic so it doesn't complicate the general case
Always look for node in this case - it'll fall back on returning everything if a full compilation unit was passed
